fix: media minor css changes and split media components
