+++ .gitignore
@@ -8,6 +8,8 @@ tmp
 # dependencies
 node_modules
 
+uploads
+
 # IDEs and editors
 /.idea
 .project
