+++ apps/frontend/src/components/launches/bot.picture.tsx
@@ -6,7 +6,7 @@ import { Input } from '@gitroom/react/form/input';
 import { Button } from '@gitroom/react/form/button';
 import { useFetch } from '@gitroom/helpers/utils/custom.fetch';
 import { useToaster } from '@gitroom/react/toaster/toaster';
-import { showMediaBox } from '@gitroom/frontend/components/media/media.component';
+import { showMediaBox } from '../media/mediaboxmodal.component';
 
 export const BotPicture: FC<{
   integration: Integrations;
